feat: labels for settings sections and options
