+++ admin/src/lib/labelsList.js
@@ -1,0 +1,15 @@
+const labelsList = {
+	paid: { name: 'Paid service', color: '#00c996' },
+	free: { name: 'Free' },
+	experimental: { name: 'Experimental', color: '#ff8875' },
+	alpha: { name: 'Alpha' },
+	beta: { name: 'Beta', color: '#75a9ff' },
+	expert: { name: 'Experts', color: '#ffc996' },
+	seo: { name: 'SEO', color: '#D4C5F9' },
+	cron: { name: 'Cron' },
+	performance: { name: 'Performance', color: '#65B5FF' },
+	tools: { name: 'Tools', color: '#FFD189' },
+	ai: { name: 'AI', color: '#ff7a7a' },
+};
+
+export default labelsList;
